Misc fixes

+++ sources_custom/mail.php
@@ -27,18 +27,18 @@
  * @range  1 5
  * @param  ?array $attachments An list of attachments (each attachment being a map, path=>filename) (null: none)
  * @param  boolean $no_cc Whether to NOT CC to the CC address
- * @param  ?MEMBER $as Convert Comcode->Tempcode as this member (a privilege thing: we don't want people being able to use admin rights by default!) (null: guest)
+ * @param  ?MEMBER $as Convert Comcode->tempcode as this member (a privilege thing: we don't want people being able to use admin rights by default!) (null: guest)
  * @param  boolean $as_admin Replace above with arbitrary admin
  * @param  boolean $in_html HTML-only
  * @param  boolean $coming_out_of_queue Whether to bypass queueing, because this code is running as a part of the queue management tools
  * @param  ID_TEXT $mail_template The template used to show the email
- * @param  boolean $bypass_queue Whether to bypass queueing
+ * @param  ?boolean $bypass_queue Whether to bypass queueing (null: auto-decide)
  * @param  ?array $extra_cc_addresses Extra CC addresses to use (null: none)
  * @param  ?array $extra_bcc_addresses Extra BCC addresses to use (null: none)
  * @param  ?TIME $require_recipient_valid_since Implement the Require-Recipient-Valid-Since header (null: no restriction)
  * @return ?Tempcode A full page (not complete XHTML) piece of Tempcode to output (null: it worked so no Tempcode message)
  */
-function mail_wrap($subject_line, $message_raw, $to_email = null, $to_name = null, $from_email = '', $from_name = '', $priority = 3, $attachments = null, $no_cc = false, $as = null, $as_admin = false, $in_html = false, $coming_out_of_queue = false, $mail_template = 'MAIL', $bypass_queue = false, $extra_cc_addresses = null, $extra_bcc_addresses = null, $require_recipient_valid_since = null)
+function mail_wrap($subject_line, $message_raw, $to_email = null, $to_name = null, $from_email = '', $from_name = '', $priority = 3, $attachments = null, $no_cc = false, $as = null, $as_admin = false, $in_html = false, $coming_out_of_queue = false, $mail_template = 'MAIL', $bypass_queue = null, $extra_cc_addresses = null, $extra_bcc_addresses = null, $require_recipient_valid_since = null)
 {
     if (get_option('smtp_sockets_use') == '0') {
         return non_overridden__mail_wrap($subject_line, $message_raw, $to_email, $to_name, $from_email, $from_name, $priority, $attachments, $no_cc, $as, $as_admin, $in_html, $coming_out_of_queue, $mail_template, $bypass_queue, $extra_cc_addresses, $extra_bcc_addresses, $require_recipient_valid_since);
@@ -65,9 +65,6 @@ function mail_wrap($subject_line, $message_raw, $to_email = null, $to_name = nul
     if (is_null($as)) {
         $as = $GLOBALS['FORUM_DRIVER']->get_guest_id();
     }
-    if (is_null($extra_cc_addresses)) {
-        $extra_cc_addresses = array();
-    }
 
     if (count($attachments) == 0) {
         $attachments = null;
@@ -80,7 +77,7 @@ function mail_wrap($subject_line, $message_raw, $to_email = null, $to_name = nul
     }
 
     if (!$coming_out_of_queue) {
-        if (!$GLOBALS['SITE_DB']->table_is_locked('logged_mail_messages')) {
+        if ((mt_rand(0, 100) == 1) && (!$GLOBALS['SITE_DB']->table_is_locked('logged_mail_messages'))) {
             $GLOBALS['SITE_DB']->query('DELETE FROM ' . get_table_prefix() . 'logged_mail_messages WHERE m_date_and_time<' . strval(time() - 60 * 60 * 24 * 14) . ' AND m_queued=0'); // Log it all for 2 weeks, then delete
         }
 
@@ -93,8 +90,12 @@ function mail_wrap($subject_line, $message_raw, $to_email = null, $to_name = nul
             }
         }
 
+        if (!$in_html) {
+            inject_web_resources_context_to_comcode($message_raw);
+        }
+
         $GLOBALS['SITE_DB']->query_insert('logged_mail_messages', array(
-            'm_subject' => $subject_line,
+            'm_subject' => cms_mb_substr($subject_line, 0, 255),
             'm_message' => $message_raw,
             'm_to_email' => serialize($to_email),
             'm_to_name' => serialize($to_name),
@@ -103,7 +104,7 @@ function mail_wrap($subject_line, $message_raw, $to_email = null, $to_name = nul
             'm_join_time' => $require_recipient_valid_since,
             'm_from_email' => $from_email,
             'm_from_name' => $from_name,
-            'm_priority' => 3,
+            'm_priority' => $priority,
             'm_attachments' => serialize($attachments),
             'm_no_cc' => $no_cc ? 1 : 0,
             'm_as' => $as,
@@ -114,7 +115,7 @@ function mail_wrap($subject_line, $message_raw, $to_email = null, $to_name = nul
             'm_url' => get_self_url(true),
             'm_queued' => $through_queue ? 1 : 0,
             'm_template' => $mail_template,
-        ));
+        ), false, !$through_queue); // No errors if we don't NEED this to work
 
         if ($through_queue) {
             return null;
@@ -167,17 +168,22 @@ function mail_wrap($subject_line, $message_raw, $to_email = null, $to_name = nul
     if ($from_name == '') {
         $from_name = get_site_name();
     }
+    $from_email = str_replace(array("\r", "\n"), array('', ''), $from_email);
+    $from_name = str_replace(array("\r", "\n"), array('', ''), $from_name);
 
-    cms_profile_start_for('mail_wrap');
+    if (!$coming_out_of_queue) {
+        cms_profile_start_for('mail_wrap');
+    }
 
     $theme = method_exists($GLOBALS['FORUM_DRIVER'], 'get_theme') ? $GLOBALS['FORUM_DRIVER']->get_theme() : 'default';
-    if ($theme == 'default') { // Sucks, probably due to sending from Admin Zone...
+    if ($theme == 'default' || $theme == 'admin') { // Sucks, probably due to sending from Admin Zone...
         $theme = $GLOBALS['FORUM_DRIVER']->get_theme(''); // ... So get theme of welcome zone
     }
 
     // Our subject
-    $_subject = do_template('MAIL_SUBJECT', array('_GUID' => '4c7eefb7296e7b7d4f3a4ef0eeeca658', 'SUBJECT_LINE' => $subject_line), $lang, false, null, '.txt', 'text', $theme);
-    $subject = $_subject->evaluate($lang); // Note that this is slightly against spec, because characters aren't forced to be printable us-ascii. But it's better we allow this (which works in practice) than risk incompatibility via charset-base64 encoding.
+    $subject = do_template('MAIL_SUBJECT', array('_GUID' => '44a57c666bb00f96723256e26aade9e5', 'SUBJECT_LINE' => $subject_line), $lang, false, null, '.txt', 'text', $theme);
+    $tightened_subject = $subject->evaluate($lang); // Note that this is slightly against spec, because characters aren't forced to be printable us-ascii. But it's better we allow this (which works in practice) than risk incompatibility via charset-base64 encoding.
+    $tightened_subject = str_replace(array("\r", "\n"), array('', ''), $tightened_subject);
 
     // Evaluate message. Needs doing early so we know if we have any headers
 
@@ -191,39 +197,64 @@ function mail_wrap($subject_line, $message_raw, $to_email = null, $to_name = nul
     global $CID_IMG_ATTACHMENT;
     $CID_IMG_ATTACHMENT = array();
 
-    // Decide message
-    $GLOBALS['NO_LINK_TITLES'] = true;
-    global $LAX_COMCODE;
-    $temp = $LAX_COMCODE;
-    $LAX_COMCODE = true;
-    $html_content = comcode_to_tempcode($message_raw, $as, $as_admin);
-    $LAX_COMCODE = $temp;
-    $GLOBALS['NO_LINK_TITLES'] = false;
+    // Evaluate message. Needs doing early so we know if we have any headers
     if (!$in_html) {
-        $_html_content = $html_content->evaluate($lang);
-        $_html_content = preg_replace('#(keep|for)_session=\w*#', 'filtered=1', $_html_content);
-        if (strpos($_html_content, '<html') !== false) {
-            $message_html = make_string_tempcode($_html_content);
-        } else {
-            $message_html = do_template($mail_template, array(
-                '_GUID' => 'b23069c20202aa59b7450ebf8d49cde1',
-                'CSS' => '{CSS}',
-                'LOGOURL' => get_logo_url(''),
-                'LANG' => $lang,
-                'TITLE' => $subject,
-                'CONTENT' => $_html_content,
-            ), $lang, false, null, '.tpl', 'templates', $theme);
-        }
-        $css = css_tempcode(true, true, $message_html->evaluate($lang), $theme);
-        $_css = $css->evaluate($lang);
-        if (get_option('allow_ext_images') != '1') {
-            $_css = preg_replace_callback('#url\(["\']?(http://[^"]*)["\']?\)#U', '_mail_css_rep_callback', $_css);
-        }
-        $html_evaluated = $message_html->evaluate($lang);
-        $html_evaluated = str_replace('{CSS}', $_css, $html_evaluated);
+        $cache_sig = serialize(array(
+            $lang,
+            $mail_template,
+            $subject,
+            $theme,
+            crc32($message_raw),
+        ));
 
-        // Cleanup the Comcode a bit
-        $message_plain = comcode_to_clean_text($message_raw);
+        static $html_content_cache = array();
+        if (isset($html_content_cache[$cache_sig])) {
+            list($html_evaluated, $message_plain, $EMAIL_ATTACHMENTS) = $html_content_cache[$cache_sig];
+        } else {
+            require_code('media_renderer');
+            push_media_mode(peek_media_mode() | MEDIA_LOWFI);
+
+            $GLOBALS['NO_LINK_TITLES'] = true;
+            global $LAX_COMCODE;
+            $temp = $LAX_COMCODE;
+            $LAX_COMCODE = true;
+            $html_content = comcode_to_tempcode($message_raw, $as, $as_admin);
+            $LAX_COMCODE = $temp;
+            $GLOBALS['NO_LINK_TITLES'] = false;
+
+            $_html_content = $html_content->evaluate($lang);
+            $_html_content = preg_replace('#(keep|for)_session=\w*#', 'filtered=1', $_html_content);
+            if (strpos($_html_content, '<html') !== false) {
+                $message_html = make_string_tempcode($_html_content);
+            } else {
+                $message_html = do_template($mail_template, array(
+                    '_GUID' => 'b23069c20202aa59b7450ebf8d49cde1',
+                    'CSS' => '{CSS}',
+                    'LOGOURL' => get_logo_url(''),
+                    'LANG' => $lang,
+                    'TITLE' => $subject,
+                    'CONTENT' => $_html_content,
+                ), $lang, false, null, '.tpl', 'templates', $theme);
+            }
+            require_css('email');
+            $css = css_tempcode(true, false, $message_html->evaluate($lang), $theme);
+            $_css = $css->evaluate($lang);
+            if (!GOOGLE_APPENGINE) {
+                if (get_option('allow_ext_images') != '1') {
+                    $_css = preg_replace_callback('#url\(["\']?(http://[^"]*)["\']?\)#U', '_mail_css_rep_callback', $_css);
+                }
+            }
+            $html_evaluated = $message_html->evaluate($lang);
+            $html_evaluated = str_replace('{CSS}', $_css, $html_evaluated);
+
+            // Cleanup the Comcode a bit
+            $message_plain = comcode_to_clean_text($message_raw);
+
+            $html_content_cache[$cache_sig] = array($html_evaluated, $message_plain, $EMAIL_ATTACHMENTS);
+
+            pop_media_mode();
+        }
+        $attachments = array_merge(is_null($attachments) ? array() : $attachments, $EMAIL_ATTACHMENTS);
     } else {
         $html_evaluated = $message_raw;
     }
@@ -234,7 +265,9 @@ function mail_wrap($subject_line, $message_raw, $to_email = null, $to_name = nul
 
     // CID attachments
     if (get_option('allow_ext_images') != '1') {
+        $cid_before = array_keys($CID_IMG_ATTACHMENT);
         $html_evaluated = preg_replace_callback('#<img\s([^>]*)src="(http://[^"]*)"#U', '_mail_img_rep_callback', $html_evaluated);
+        $cid_just_html = array_diff(array_keys($CID_IMG_ATTACHMENT), $cid_before);
         $matches = array();
         foreach (array('#<([^"<>]*\s)style="([^"]*)"#', '#<style( [^<>]*)?' . '>(.*)</style>#Us') as $over) {
             $num_matches = preg_match_all($over, $html_evaluated, $matches);
@@ -245,6 +278,12 @@ function mail_wrap($subject_line, $message_raw, $to_email = null, $to_name = nul
                     $html_evaluated = str_replace($matches[0][$i], $altered_outer, $html_evaluated);
                 }
             }
+        }
+
+        // This is a hack to stop the images used by CSS showing as attachments in some mail clients
+        $cid_just_css = array_diff(array_keys($CID_IMG_ATTACHMENT), $cid_just_html);
+        foreach ($cid_just_css as $cid) {
+            $html_evaluated .= '<img width="0" height="0" src="cid:' . $cid . '" />';
         }
     }
     $cid_attachments = array();
@@ -429,10 +468,12 @@ function mail_wrap($subject_line, $message_raw, $to_email = null, $to_name = nul
 
     // Send the message, and error collection
     $error = '';
+    $worked = true;
     try {
         $result = $mailer->send($message);
     } catch (Exception $e) {
         $error = $e->getMessage();
+        $worked = false;
     }
     if (($error == '') && (!$result)) {
         $error = 'Unknown error';
@@ -450,13 +491,16 @@ function mail_wrap($subject_line, $message_raw, $to_email = null, $to_name = nul
         }
     }
 
-    cms_profile_end_for('mail_wrap', $subject_line);
+    if (!$coming_out_of_queue) {
+        cms_profile_end_for('mail_wrap', $subject_line);
+    }
 
-    // Return / Error handling
-    $SENDING_MAIL = false;
-    if ($error != '') {
+    if (!$worked) {
+        $SENDING_MAIL = false;
         require_code('site');
         attach_message(!is_null($error) ? make_string_tempcode($error) : do_lang_tempcode('MAIL_FAIL', escape_html(get_option('staff_address'))), 'warn');
     }
+
+    $SENDING_MAIL = false;
     return null;
 }
